backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
httpx>=0.24.0
python-multipart>=0.0.5 